로그인 정보 수정, 검색 오류 수정, 페이징 구현

+++ src/main/java/com/fastquick/service/OrderService.java
@@ -59,7 +59,7 @@ public class OrderService {
 	}
 
 	@Transactional
-	public void updateOrderById (Long id){
+	public void updateOrderById (Integer id){
 		Optional<Member> member = memberRepository.findById(id);
 		saveOrderByMember(member.get());
 	}
